abstract command provider tests since we should have consistent behavior across all command providers
