+++ Trelnex.Core.Data.Tests/InternalsVisibleTo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("ProxyBuilder")]
+[assembly: InternalsVisibleTo("Trelnex.Core.Api.Tests")]
